* feat(install.sh): add service_status function to check the status of a service * feat(install.sh): add restart_service function to restart a service * feat(install.sh): add download_release_file function to download a release file from a GitHub repository * feat(install.sh): add register_icefiresql_service function to register the icefiresql service * feat(install.sh): add start_service function to start a service * feat(install.sh): add stop_service function to stop a service * feat(install.sh): add download_tarball_binary function to download a tarball binary from a GitHub repository * feat(install.sh): add register_sonrd_service function to register the sonrd service
* feat(install.sh): add support for downloading and installing IceFireDB SQL binary
* feat(install.sh): create directory for IceFireDB SQL data storage
* feat(install.sh): download and install IceFireDB SQL configuration file
* feat(install.sh): add support for registering IceFireDB SQL service
* feat(install.sh): add support for upgrading IceFireDB and Sonr
* feat(install.sh): restart IceFireDB SQL service during upgrade
* feat(install.sh): restart IceFireDB Key-Value service during upgrade
* feat(install.sh): restart Sonr service during upgrade
* feat(install.sh): add option to check status of Sonr service

+++ scripts/install.sh
@@ -4,20 +4,57 @@
 #// ! ||--------------------------------------------------------------------------------||
 #// ! ||                                    Utilities                                   ||
 #// ! ||--------------------------------------------------------------------------------||
+
+service_status() {
+    if [ "$(uname)" == "Linux" ]; then
+        sudo systemctl status $1
+    else
+        echo "This function can only be run on Linux."
+        exit 1
+    fi
+}
+
+restart_service() {
+    if [ "$(uname)" == "Linux" ]; then
+        sudo systemctl daemon-reload
+        if systemctl is-enabled --quiet $1; then
+            sudo systemctl stop $1
+            sudo systemctl restart $1
+        fi
+    else
+        echo "This function can only be run on Linux."
+        exit 1
+    fi
+}
+
 start_service() {
-    sudo systemctl daemon-reload
-    if ! systemctl is-enabled --quiet $1; then
-        sudo systemctl enable $1
-        sudo systemctl start $1
+    if [ "$(uname)" == "Linux" ]; then
+        if ! systemctl is-enabled --quiet $1; then
+            sudo systemctl enable $1
+            sudo systemctl start $1
+        fi
+    else
+        echo "This function can only be run on Linux."
+        exit 1
     fi
 }
 
 stop_service() {
-    sudo systemctl daemon-reload
-    if systemctl is-enabled --quiet $1; then
-        sudo systemctl stop $1
-        sudo systemctl disable $1
+    if [ "$(uname)" == "Linux" ]; then
+        if systemctl is-enabled --quiet $1; then
+            sudo systemctl stop $1
+        fi
+    else
+        echo "This function can only be run on Linux."
+        exit 1
     fi
+}
+
+download_release_file() {
+    REPO=$1
+    FILE=$2
+    OUTPUT=$3
+    wget https://github.com/$REPO/releases/latest/download/$FILE -O $OUTPUT
 }
 
 download_tarball_binary() {
@@ -55,6 +92,26 @@ EOF
 start_service icefirekv
 }
 
+# Register icefiresql service
+register_icefiresql_service() {
+# Setup systemd for IceFireDB
+sudo cat << EOF > /etc/systemd/system/icefiresql.service
+[Unit]
+Description=IceFireSQLite Service
+After=network-online.target
+
+[Service]
+User=root
+ExecStart=/usr/local/bin/icefiresql -c /var/lib/icefiresql/config.yml
+LimitNOFILE=4096
+
+[Install]
+WantedBy=multi-user.target
+EOF
+
+start_service icefiresql
+}
+
 # Setup systemd for sonrd
 register_sonrd_service() {
 sudo cat << EOF > /etc/systemd/system/sonrd.service
@@ -86,49 +143,55 @@ start_service sonrd
 install() {
     download_tarball_binary sonr-io/sonr sonrd
     download_tarball_binary sonr-io/IceFireDB icefirekv
+    download_tarball_binary sonr-io/IceFireDB icefiresql
+    mkdir -p /var/lib/icefiresql
+    download_release_file sonr-io/IceFireDB config.sqlite.yml /var/lib/icefiresql/config.yml
 }
 
 register_services() {
     if [ "$(uname)" == "Linux" ]; then
         register_icefirekv_service
+        register_icefiresql_service
         register_sonrd_service
     else
         echo "This function can only be run on Linux."
+        exit 1
     fi
 }
 
 upgrade() {
-    stop_service sonrd
-    stop_service icefirekv
     download_tarball_binary sonr-io/sonr sonrd
     download_tarball_binary sonr-io/IceFireDB icefirekv
-    start_service icefirekv
-    start_service sonrd
+    download_tarball_binary sonr-io/IceFireDB icefiresql
+    restart_service icefirekv
+    restart_service icefiresql
+    restart_service sonrd
 }
 
 #// ! ||--------------------------------------------------------------------------------||
 #// ! ||                                    Startup                                     ||
 #// ! ||--------------------------------------------------------------------------------||
 
-OPTIONS=("Install Sonr and Deps" "Register System Service" "Upgrade to latest" "Check status of Service" "Exit")
+OPTIONS=("Install Sonr and Deps" "Register Systemd Service" "Upgrade Installation" "Check Service Status" "Exit")
 PS3='Please enter your choice: '
 
 select CHOICE in "${OPTIONS[@]}"
 do
+  echo ''
   case $CHOICE in
     "Install Sonr and Deps")
         install
         break
         ;;
-    "Register System Service")
+    "Register Systemd Service")
         register_services
         break
         ;;
-    "Upgrade to latest")
+    "Upgrade Installation")
         upgrade
         break
         ;;
-    "Check status of Service")
+    "Check Service Status")
         sudo systemctl status sonrd
         break
         ;;
